backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
peewee==3.17.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.4.0

